+++ src/tree/tree_manager/canonical.rs
@@ -1,88 +1,22 @@
 use std::collections::HashMap;
-use std::marker::PhantomData;
 use std::sync::Arc;
 
-use ethers::abi::{AbiDecode, RawLog};
+use ethers::abi::AbiDecode;
 use ethers::contract::{EthCall, EthEvent};
 use ethers::providers::Middleware;
-use ethers::types::{
-    Filter, Log, Selector, Transaction, ValueOrArray, H160, H256, U256,
-};
+use ethers::types::{Log, Selector, Transaction, H256, U256};
 use futures::{StreamExt, TryStreamExt};
 use tokio::sync::mpsc::Sender;
 use tokio::task::JoinHandle;
 
-use super::block_scanner::BlockScanner;
-use super::error::{WorldTreeError, WorldTreeResult};
-use super::identity_tree::LeafUpdates;
-use super::{Hash, LeafIndex};
+use super::TreeVersion;
 use crate::abi::{
-    DeleteIdentitiesCall, RegisterIdentitiesCall, RootAddedFilter,
-    TreeChangedFilter,
+    DeleteIdentitiesCall, RegisterIdentitiesCall, TreeChangedFilter,
 };
-
-pub const BLOCK_SCANNER_SLEEP_TIME: u64 = 5;
-
-pub trait TreeVersion: Default {
-    type ChannelData;
-
-    fn spawn<M: Middleware + 'static>(
-        tx: Sender<Self::ChannelData>,
-        block_scanner: Arc<BlockScanner<M>>,
-    ) -> JoinHandle<WorldTreeResult<()>>;
-
-    fn tree_changed_signature() -> H256;
-}
-
-pub struct TreeManager<M: Middleware + 'static, T: TreeVersion> {
-    pub address: H160,
-    pub block_scanner: Arc<BlockScanner<M>>,
-    pub chain_id: u64,
-    _tree_version: PhantomData<T>,
-}
-
-impl<M, T> TreeManager<M, T>
-where
-    M: Middleware + 'static,
-    T: TreeVersion,
-{
-    pub async fn new(
-        address: H160,
-        window_size: u64,
-        last_synced_block: u64,
-        middleware: Arc<M>,
-    ) -> WorldTreeResult<Self> {
-        let chain_id = middleware.get_chainid().await?.as_u64();
-
-        let filter = Filter::new()
-            .address(address)
-            .topic0(ValueOrArray::Value(T::tree_changed_signature()));
-
-        let block_scanner = Arc::new(
-            BlockScanner::new(
-                middleware,
-                window_size,
-                last_synced_block,
-                filter,
-            )
-            .await?,
-        );
-
-        Ok(Self {
-            address,
-            block_scanner,
-            chain_id,
-            _tree_version: PhantomData,
-        })
-    }
-
-    pub fn spawn(
-        &self,
-        tx: Sender<T::ChannelData>,
-    ) -> JoinHandle<WorldTreeResult<()>> {
-        T::spawn(tx, self.block_scanner.clone())
-    }
-}
+use crate::tree::block_scanner::BlockScanner;
+use crate::tree::error::{WorldTreeError, WorldTreeResult};
+use crate::tree::identity_tree::LeafUpdates;
+use crate::tree::{Hash, LeafIndex};
 
 #[derive(Default)]
 pub struct CanonicalTree;
@@ -135,55 +69,6 @@ impl TreeVersion for CanonicalTree {
 
     fn tree_changed_signature() -> H256 {
         TreeChangedFilter::signature()
-    }
-}
-
-#[derive(Default)]
-pub struct BridgedTree;
-
-#[derive(Debug, Clone)]
-pub struct BridgeTreeUpdate {
-    pub chain_id: u64,
-    pub root: Hash,
-}
-
-impl TreeVersion for BridgedTree {
-    type ChannelData = BridgeTreeUpdate;
-
-    fn spawn<M: Middleware + 'static>(
-        tx: Sender<Self::ChannelData>,
-        block_scanner: Arc<BlockScanner<M>>,
-    ) -> JoinHandle<WorldTreeResult<()>> {
-        tokio::spawn(async move {
-            let chain_id =
-                block_scanner.middleware.get_chainid().await?.as_u64();
-
-            tracing::info!(?chain_id, "Starting bridged tree manager");
-            block_scanner
-                .block_stream()
-                // Retrieve logs concurrently
-                // Setting this too high can cause a 502
-                .buffered(10)
-                // Process logs sequentially
-                .try_for_each(|logs| async {
-                    for log in logs {
-                        // Extract the root from the RootAdded log
-                        let data =
-                            RootAddedFilter::decode_log(&RawLog::from(log))?;
-                        let root = Hash::from_limbs(data.root.0);
-
-                        tracing::info!(?chain_id, ?root, "Root updated");
-                        tx.send(BridgeTreeUpdate { chain_id, root }).await?;
-                    }
-                    Ok(())
-                })
-                .await?;
-            Ok(())
-        })
-    }
-
-    fn tree_changed_signature() -> H256 {
-        RootAddedFilter::signature()
     }
 }
